Removed ProvingSystemUtils trait
